Version bump to 2.2.0-volders
We need to enable the source map generation but we got some erros
on ember-cli-uglify. So we had to create our own version to fix it. More
information here:

  https://github.com/vertragswecker/ember-cli-uglify/commit/2acd034252e26c7d9ba909bcefe5cead7eb129f9

Then, we start having this similar error on broccoli-uglify-sourcemap
while building the app for production:

```
node_modules/ember-cli-uglify/node_modules/broccoli-uglify-sourcemap/index.js:3
const walkSync = require('walk-sync');
^^^^^
SyntaxError: Use of const in strict mode.
```

This is because we are using an old version of node that not deal with
new javascript syntax very well.

So, to fix the issue without the need to the node version right
now, we created this version with this commit on top of the
version 2.2.0. The we have to use this version as dependency on
ember-cli-uglify.